added support for receiving message data that is longer than the buffer.
The buffer must still be large enough to hold the header, topic and
some of the message data. The first part of the data will be delivered
as normal via the MQTT_EVENT_TYPE_PUBLISH, extra data beyond the end
of the buffer will be delivered via a new
MQTT_EVENT_TYPE_PUBLISH_CONTINUATION event.